Split logic into separate files and namespace vscode
